test: split idle forecast coverage
